Plus Minus
Given an array of integers, calculate which fraction of its elements are positive, which fraction of its elements are negative, and which fraction of its elements are zeros. Print the decimal value of each fraction on a new line.

Input Format
The first line contains an integer, n, denoting the size of the array. 
The second line contains n space-separated integers describing an array of numbers 
arr(a0,a1,a2,......a(n-1))

Output Format
You must print the following 3 lines:
A decimal representing of the fraction of positive numbers in the array compared to its size.
A decimal representing of the fraction of negative numbers in the array compared to its size.
A decimal representing of the fraction of zeros in the array compared to its size.

Sample Input
6
-4 3 -9 0 4 1         

Sample Output
0.500000
0.333333
0.166667

Explanation
There are 3 positive numbers, 2 negative numbers, and 1 zero in the array. 
The proportions of occurrence are positive: 3/6 negative: 2/6 and zeros:1/6