Grading Students
At Harvard University, a passing grade is any grade 40 points or higher on a 100 point scale. Sam is a professor at the university and likes to round each student’s grade according to the following rules:
If the difference between the grade and the next higher multiple of 5 is less than 3, round to the next higher multiple of 5
If the grade is less than 38, don’t bother as it’s still a failing grade
Automate the rounding process then round a list of grades and print the results.
Input Format
The first line contains a single integer denoting n (the number of students). 
Each line i of the n subsequent lines contains a single integer,grades , denoting student i's grade.
Output Format
For each grade of the n grades, print the rounded grade on a new line.
Sample Input 0
4
73
67
38
33
Sample Output 0
75
67
40
33
Explanation 0
The first grade, 73 is two below the next higher multiple of 5, so it rounds to . 75
63 is 3 points less than the next higher multiple of 5 so it doesn’t round. 
38, like 73, rounds up to next higher multiple of 5, or 40 in this case. 
33 is less than ,38 so it does not round.