Between Two Sets
You will be given two arrays of integers. You will be asked to determine all integers that satisfy the following two conditions:
The elements of the first array are all factors of the integer being considered
The integer being considered is a factor of all elements of the second array
These numbers are referred to as being between the two arrays. You must determine how many such numbers exist.
Input Format
The first line contains two space-separated integers describing the respective values of n, the number of elements in array a, and m, the number of elements in array b. 
The second line contains n distinct space-separated integers describing . a0,a1,....an-1 
The third line contains m distinct space-separated integers describing .b0,b1,....bn-1 
Output Format
Print the number of integers that are considered to be between a and b.
Sample Input
2 3
2 4
16 32 96
Sample Output
3
Explanation
2 and 4 divide evenly into 4, 8, 12 and 16. 
4, 8 and 16 divide evenly into 16, 32, 96.
4, 8 and 16 are the only three numbers for which each element of A is a factor and each is a factor of all elements of B.