Forming a Magic Square
We define a magic square to be an nxn matrix of distinct positive integers from 1 to n^2 where the sum of any row, column, or diagonal (of length n) is always equal to the same number (i.e., the magic constant).
Consider a 3x3 matrix, s, of integers in the inclusive range [1,9]. We can convert any digit, a, to any other digit, b, in the range [1,9] at cost |a-b|.
Given s, convert it into a magic square at minimal cost by changing zero or more of its digits. Then print this cost on a new line.
Input Format
There are 3 lines of input. Each line describes a row of the matrix in the form of 3 space-separated integers denoting the respective first, second, and third elements of that row.
Output Format
Print an integer denoting the minimum cost of turning matrix s into a magic square.
Sample Input 0
4 9 2
3 5 7
8 1 5
Sample Output 0
1
Sample Input 1
4 8 2
4 5 7
6 1 6
Sample Output 1
4