Jumping on the Clouds: Revisited
Aerith is playing a cloud hopping game. In this game, there are sequentially numbered clouds that can be thunderheads or cumulus clouds. Her character must jump from cloud to cloud until it reaches the start again.

To play, Aerith is given an array of clouds, c and an energy level c=100. She starts from c[0] and uses 1 unit of energy to make a jump of size k to cloud c[(i+k)%n]. If Aerith lands on a thundercloud, c[i]=1, her energy (E) decreases by 2 additional units. The game ends when Aerith lands back on cloud 0.

Given the values of n, k, and the configuration of the clouds as an array c, can you determine the final value of c after the game ends?

Input Format

The first line contains two space-separated integers, n and k, the number of clouds and the jump distance. 
The second line contains n space-separated integers c[i] . Each cloud is described as follows:

If c[i]=0, then cloud i is a cumulus cloud.
If c[i]=1, then cloud i is a thunderhead.

Output Format

Print the final value of E on a new line.

Sample Input

8 2
0 0 1 0 0 1 1 0

Sample Output

92