Find Digits
An integer d is a divisor of an integer n if the remainder of n/d=0.

Given an integer, for each digit that makes up the integer determine whether it is a divisor. Count the number of divisors occurring within the integer.

Input Format

The first line is an integer, t, indicating the number of test cases. 
The t subsequent lines each contain an integer, n.

Output Format

For every test case, count the number of digits in n that are divisors of n. Print each answer on a new line.

Sample Input

2
12
1012

Sample Output

2
3