Sherlock and Squares
Watson likes to challenge Sherlock's math ability. He will provide a starting and ending value describing a range of integers. Sherlock must determine the number of square integers within that range, inclusive of the endpoints.

Input Format

The first line contains T, the number of test cases. 
Each of the next T lines contains two space-separated integers denoting A and B, the starting and ending integers in the ranges.

Output Format

For each test case, print the number of square integers in the range on a new line.

Sample Input

2
3 9
17 24
Sample Output

2
0