Library Fine
Your local library needs your help! Given the expected and actual return dates for a library book, create a program that calculates the fine (if any). The fee structure is as follows:
1.) If the book is returned on or before the expected return date, no fine will be charged 
2.) If the book is returned after the expected return day but still within the same calendar month and year as the expected return date, file = 15 x No. of days late 
3.) If the book is returned after the expected return month but still within the same calendar year as the expected return date, the fine = 500 x no. of months late 
4.) If the book is returned after the calendar year in which it was expected, there is a fixed fine of 10000 

Input Format

The first line contains 3 space-separated integers denoting the respective days, month, and year on which the book was returned. 
The second line contains 3 space-separated integers denoting the respective days, month, and year on which the book was due to be returned.

Output Format

Print a single integer denoting the library fine for the book received as input.

Sample Input

9 6 2015
6 6 2015

Sample Output

45