Flatland Space Stations
Flatland is a country with a number of cities, some of which have space stations. Cities are numbered consecutively and each has a road of  length 1km connecting it to the next city. It is not a circular route, so the first city doesn't connect with the last city. Determine the maximum distance from any city to it's nearest space station.

Input Format

The first line consists of two space-separated integers, n and m. 
The second line contains m space-separated integers, the indices of each city having a space-station. These values are unordered and unique.

Output Format

Print an integer denoting the maximum distance that an astronaut in a Flatland city would need to travel to reach the nearest space station.

Sample Input 0

5 2
0 4
Sample Output 0

2

Sample Input 1

6 6
0 1 2 4 3 5
Sample Output 1

0